requirements.txt
Flask==3.1.2
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
